refactor: move deriveWorldStateForClient function to separate module
